requirements.txt
deepgram-sdk>=5.0.0
python-dotenv>=1.0.0
